+++ src/parser/block/paragraph/hashtag.test.ts
@@ -16,14 +16,15 @@ describe('Unit: parser/block/paragraph/hashtag', () => {
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('#0')), [['<a class="hashtag" rel="noopener">#0</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('#a')), [['<a class="hashtag" rel="noopener">#a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('#A')), [['<a class="hashtag" rel="noopener">#A</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('#a ')), [['<a class="hashtag" rel="noopener">#a</a>'], ' ']);
-      assert.deepStrictEqual(inspect(parser('#a\n')), [['<a class="hashtag" rel="noopener">#a</a>'], '\n']);
-      assert.deepStrictEqual(inspect(parser('#\\')), [['<a class="hashtag" rel="noopener">#\\</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('#\\ ')), [['<a class="hashtag" rel="noopener">#\\</a>'], ' ']);
-      assert.deepStrictEqual(inspect(parser('#\\\n')), [['<a class="hashtag" rel="noopener">#\\</a>'], '\n']);
+      assert.deepStrictEqual(inspect(parser('#0')), [['<a class="hashtag" rel="noopener" data-level="1">#0</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#a')), [['<a class="hashtag" rel="noopener" data-level="1">#a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#A')), [['<a class="hashtag" rel="noopener" data-level="1">#A</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#a ')), [['<a class="hashtag" rel="noopener" data-level="1">#a</a>'], ' ']);
+      assert.deepStrictEqual(inspect(parser('#a\n')), [['<a class="hashtag" rel="noopener" data-level="1">#a</a>'], '\n']);
+      assert.deepStrictEqual(inspect(parser('#\\')), [['<a class="hashtag" rel="noopener" data-level="1">#\\</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#\\ ')), [['<a class="hashtag" rel="noopener" data-level="1">#\\</a>'], ' ']);
+      assert.deepStrictEqual(inspect(parser('#\\\n')), [['<a class="hashtag" rel="noopener" data-level="1">#\\</a>'], '\n']);
+      assert.deepStrictEqual(inspect(parser('##0')), [['<a class="hashtag" rel="noopener" data-level="2">##0</a>'], '']);
     });
 
   });