+++ src/parser/block/paragraph/hashtag.ts
@@ -1,13 +1,17 @@
 ﻿import { ParagraphParser } from '../../block';
-import { union, some, surround, fmap } from '../../../combinator';
+import { union, some, capture, fmap } from '../../../combinator';
 import { line } from '../../source/line';
 import { unescsource } from '../../source/unescapable';
 import { compress } from '../../util';
 import { html } from 'typed-dom';
 
 export const hashtag: ParagraphParser.HashtagParser = line(
-  fmap<ParagraphParser.HashtagParser>(
-    surround(/^(?=#\S)/, compress(some(union([unescsource]), /^\s/)), ''),
-    ts =>
-      [html('a', { class: 'hashtag', rel: 'noopener' }, ts)]),
+  capture(
+    /^(?=(#+)\S)/,
+    ([, { length: level }], rest) =>
+      fmap(
+        compress(some(union([unescsource]), /^\s/)),
+        ts =>
+          [html('a', { class: 'hashtag', rel: 'noopener', 'data-level': `${level}` }, ts)])
+        (rest)),
   false);